fix z80
